OI Safe: moved almost all access of the database to a new class Passwords. This new class in turn interacts with DBHelper.   It retains all PassEntry and CategoryEntry data in memory.  Encrypted data is only decrypted as needed.   Decrypted data is kept in memory.  This serves as a cache and results in a dramatic increase in speed on the CategoryList and PassList activities upon repeat use.
All references to DBHelper for passwords or categories have been 
removed from all activities and centralized inside of Passwords.  The
only remaining DBHelper calls are for the Salt and MasterPassword.

With this new class we should be able to more easily implement a
search activity.

Fixed an assortment of minor flaws surround orientation changes.
Possibly fixed I210.  Fixed I206 and I209.

This is a major overhaul.  Please help test for new bugs.


git-svn-id: https://openintents.googlecode.com/svn/trunk@1967 72b678ce-9140-0410-bee8-679b907dd61a