OI FM: Added primitive support for thumbnails. After loading the contents of a directory, a separate thread will go through each file and try to load it as a bitmap (scaled down to fit a predetermined thumbnail size).
Could use more polishing; there needs to be more padding, and it would be nice if the images were cached.

KEEP IN MIND that the debugger's exception handling causes memory leaks, so handling too many bitmaps with the debugger attached can cause OOM failures. This is a debugger-only problem. Since the thumbnail loader scales the images down while loading them, it is typically still possible to have prolonged debugging sessions.


git-svn-id: https://openintents.googlecode.com/svn/trunk@2284 72b678ce-9140-0410-bee8-679b907dd61a