OI Shopping List: Adjust classpath.
Based on http://stackoverflow.com/a/8437996/589306 and the linked
information, it seems that the _src links are the /old/ way to do
library projects, and the jars in the build path are the /new/ way.

So, instead of repeatedly removing the .jars from the eclipse build
path dialog each time I restart eclipse, it seems we should be taking
the _src links out of the classpath.